Added a bunch of HTML, implemented a leaflet map on the homepage, as well as a good deal of backend to deliver markers representing each toilet on to the leaflet map.  Still a lot to do before it's in a 'functional' state.
To Do:
As it is, toilet CharFields cannot accept single
quotes.  This may be due to character formatting
or the JSON renderer.  It might be solved by
validators, but I need to look in to those more.
Note:  Haven't tested double quotes either.

Implement a rating page that accepts input from
users.  I need to also modify the Toilet model
to include a counter of the number of times it
has been rated.  The rating field will be
increased by each rating as the rating count
field is incremented, allowing an 'average'
rating to be obtained.

I'll eventually need to refactor code to make
it more effecient.

Of course, everything needs to look better,
CSS + HTML.