Changed the way the JSON is processed in the html so that character fields in it can now use single quotes without being invalid.  Also added styling to the index page.  I also created an 'add toilet' button which currently does nothing.
To Do:
Need to learn more about HTML/CSS/JS in general,
since there seem to be some errors when
interacting with the map prior to it fully
loading.

Figure out how and what exactly the detail page
will display.

Figure out a way for the user to add toilets to
the map.